+++ tests/testthat/test-lnt_lookup.R
@@ -54,3 +54,12 @@ test_that("Lookup stat. computing in sample", {
                     `9` = c("statis", "statis", "statis", "statis", "statis", "statis"),
                     `10` = NULL))
 })
+
+test_that("Lookup stat. computing in character string", {
+  expect_equal({
+    test <- c("This contains statistical computing",
+              "This does not")
+    names(test) <- 1:2
+    lnt_lookup(test, "statistical computing", verbose = FALSE)
+    }, list(`1` = "statistical computing", `2` = NULL))
+})